Add a MLIR C Target
This commit introduces an alternative MLIR Target to llvmir, using C.
It relies on xtc-translate for the code generation from MLIR, with is packaged
in XTC-MLIR.

This target can be selected with the "target" kwarg at the creation of a
MlirCompiler, or by setting the "MLIR_TARGET" env variable.

Vectorization and OpenMP dialects are not supported yet, so many tests
are marked as XFAIL when using this new target. A scalar version of
backends/test_matmul_mlir.py has also been added.